+++ src/pages/Account/ChangelogView.php
@@ -3,7 +3,7 @@
 namespace Smr\Pages\Account;
 
 use Smr\Account;
-use Smr\Database;
+use Smr\Changelog;
 use Smr\Page\AccountPage;
 use Smr\Page\ReusableTrait;
 use Smr\Template;
@@ -26,48 +26,10 @@ class ChangelogView extends AccountPage {
 			$template->assign('ContinueHREF', $container->href());
 		}
 
-		$db = Database::getInstance();
-
-		$dbResult = $db->read('SELECT *
-					FROM version
-					WHERE went_live > :last_login
-					ORDER BY version_id DESC', [
-			'last_login' => $db->escapeNumber($this->lastLogin ?? 0),
-		]);
-
-		$versions = [];
-		foreach ($dbResult->records() as $dbRecord) {
-			$version_id = $dbRecord->getInt('version_id');
-			$version = $dbRecord->getInt('major_version') . '.' . $dbRecord->getInt('minor_version') . '.' . $dbRecord->getInt('patch_level');
-			$went_live = $dbRecord->getInt('went_live');
-
-			// get human readable format for date
-			if ($went_live > 0) {
-				$went_live = date($account->getDateTimeFormat(), $went_live);
-			} else {
-				$went_live = 'never';
-			}
-
-			$dbResult2 = $db->read('SELECT *
-						FROM changelog
-						WHERE version_id = :version_id
-						ORDER BY changelog_id', [
-				'version_id' => $db->escapeNumber($version_id),
-			]);
-			$changes = [];
-			foreach ($dbResult2->records() as $dbRecord2) {
-				$changes[] = [
-					'title' => $dbRecord2->getString('change_title'),
-					'message' => $dbRecord2->getString('change_message'),
-				];
-			}
-
-			$versions[] = [
-				'version' => $version,
-				'went_live' => $went_live,
-				'changes' => $changes,
-			];
-		}
+		$versions = Changelog::getDisplayVersions(
+			since: $this->lastLogin ?? 0,
+			dateFormat: $account->getDateTimeFormat(),
+		);
 		$template->assign('Versions', $versions);
 	}
